BTL_TDH/requirements.txt
selenium
webdriver-manager
pytesseract
schedule
Pillow
